Update intel-to-m1.md
This is just a simple update to the instructions based on a newer Expo's blog post: https://blog.expo.dev/m1-workers-on-eas-build-dcaa2c1333ad#ad50
```The instructions in previous version of this post said to use the “m1-medium” resource class. This has been replaced by the “m-medium” resource class, which more accurately describes the M-series of builders (M1, M2). The “m1-medium” resource class will continue to work but “m-medium” is now recommended.```